Allow multiple BindToPrefix sockets
When multiple BindToPrefix prefixes are configured via systemd
there will be duplicate INADDR_ANY+port combinations, which
fails to bind properly because of kresd's use of map based on
addr+port key.

This patches disables fetching of map index matching addr+port
and always create a new endpoint. The output of net.list() will
not be correct when BindToPrefix is used